requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
requests
python-multipart
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
google-generativeai
python-jose
pytest
httpx
loguru
fpdf
pypdf
pypdf2
python-magic
requests
streamlit
watchdog
torch>=2.0.0
sentence-transformers>=2.2.2
chromadb>=0.3.0
langchain>=0.0.200